requirements.txt
discord.py
fortnitepy
crayons
pymongo
requests
coloredlogs
discord-components